Add the info dictionary part of the .torrent file.
Change DecodeString to return a byte array instead of the converted string.
Make changes to all code relying on the byte array being a string.
Add a reference to a downloaded file to test whether the first item in the pieces list is the same as the computed SHA1 hash of the downloaded file.